Implemented gfal operations + renamed sh_call to ps_call
