+++ includeCMSSWlibs.py
@@ -1,4 +1,4 @@
-from RunKit.sh_tools import sh_call
+from RunKit.run_tools import ps_call
 import os
 import re
 import ROOT
@@ -6,7 +6,7 @@ pattern= "=|\n"
 def includeLibTool(tool="", wantLib=False):
     command = ["scram", "tool", "info", tool ]
     directory = os.environ['CMSSW_BASE']
-    returncode, output, err= sh_call(command, catch_stdout=True, cwd=directory, verbose=0)
+    returncode, output, err= ps_call(command, catch_stdout=True, cwd=directory, verbose=0)
     result = re.split(pattern, output)
     include_path = result[result.index("INCLUDE")+1]
     ROOT.gInterpreter.AddIncludePath(include_path)
